feat: Update draft status for blog posts
This commit updates the draft status for three blog posts: "Hello World", "My Second Post!", and "New Beginnings". The `isDraft` property is set to `true` for these posts in the Markdown frontmatter, indicating that they are still in draft mode and should not be published. This change ensures that the draft posts are properly handled and not included in the published blog pages.

Recent repository commits:
- feat: Refactor slug creation for Markdown and Sanity posts
- feat: add sanity source
- feat: upgrade dependencies
- edit: upgrade gatsby starter